Update 1034 有理数四则运算 (20分).cpp

+++ Basic Level/1034 有理数四则运算 (20分).cpp
@@ -1,17 +1,17 @@
 #include <iostream>
 using namespace std;
-long long a, b, c, d;
-long long gcd(long long a, long long b) {
+long a, b, c, d;
+long gcd(long a, long b) {
 	return !b ? a : gcd(b, a % b);
 }
-void f(long long n, long long d) {
+void f(long n, long d) {
 	if (n * d == 0) {
 		cout << (d == 0 ? "Inf" : "0");
 		return ;
 	}
 	bool flag = (n < 0 && d > 0) || (n > 0 && d < 0);
 	n = abs(n); d = abs(d);
-	long long q = n / d;
+	long q = n / d;
 	cout << (flag ? "(-" : "");
 	if (q != 0) cout << q;
 	if (n % d == 0) {
@@ -20,12 +20,12 @@ void f(long long n, long long d) {
 	}
 	if (q != 0) cout << ' ';
 	n -= q * d;
-	long long t = gcd(n, d);
+	long t = gcd(n, d);
 	n /= t; d /= t;
 	cout << n << '/' << d << (flag ? ")" : "");
 }
 int main() {
-	scanf("%lld/%lld %lld/%lld", &a, &b, &c, &d);
+	scanf("%ld/%ld %ld/%ld", &a, &b, &c, &d);
 	f(a, b); cout << " + "; f(c, d); cout << " = "; f(a*d + b*c, b*d); cout << endl;
 	f(a, b); cout << " - "; f(c, d); cout << " = "; f(a*d - b*c, b*d); cout << endl;
 	f(a, b); cout << " * "; f(c, d); cout << " = "; f(a*c, b*d); cout << endl;
